fix(types): improve typescript support for dynamic imports (#179)
