+++ lbh/components/lbh-back-to-top/back-to-top.d.ts
@@ -1,0 +1,10 @@
+class BackToTop {
+  $module: HTMLElement;
+
+  constructor($module: HTMLElement);
+
+  init(): void;
+  checkScrollPosition($element: HTMLElement): void;
+}
+
+export default BackToTop;
