+++ lbh/components/lbh-button/button.d.ts
@@ -1,0 +1,12 @@
+class Button {
+  $module: HTMLElement;
+  debounceFormSubmitTimer: number | null;
+
+  constructor($module: HTMLElement);
+
+  handleKeyDown(event: KeyboardEvent): void;
+  debounce(event: MouseEvent): void | false;
+  init(): void;
+}
+
+export default Button;
