+++ lbh/components/lbh-collapsible/collapsible.d.ts
@@ -1,0 +1,9 @@
+class Collapsible {
+  $module: HTMLElement;
+
+  constructor($module: HTMLElement);
+
+  init(): void;
+}
+
+export default Collapsible;
