+++ lbh/components/lbh-cookie-banner/cookie-banner.d.ts
@@ -1,0 +1,13 @@
+class CookieBanner {
+  $module: HTMLElement;
+  $button: HTMLElement | null;
+
+  constructor($module: HTMLElement);
+
+  init(): void;
+  checkCookies(): string | null;
+  showCookieBanner(): void;
+  bindAcceptButton(): void;
+}
+
+export default CookieBanner;
