+++ lbh/components/lbh-details/details.d.ts
@@ -1,0 +1,12 @@
+class Details {
+  $module: HTMLElement;
+
+  constructor($module: HTMLElement);
+
+  init(): void;
+  polyfillDetails(): void;
+  polyfillSetAttributes(): true;
+  polyfillHandleInputs(node: HTMLElement, callback: Function): void;
+}
+
+export default Details;
